Minor fixes and improvements (#217)
* mDNS: broadcast firmware version with avahi

https://github.com/shiftdevices/bitbox-base-internal/issues/275

Because:

* On network discovery, the BitBox App needs to be aware of the firmware version of
  the Base in case there are incompabilities regarding API endpoints (esp. a different
  pairing mechanism)
* Ideas included creating unidentified API endpoints /version or /info, but that adds
  NGINX as a webserver on port 80 and a different API technology, contradicting the
  idea of exposing as little ports as possible
* Additionally, the unauthenticated RPC call for pairing could provide version
  information as well.

This commit:

* replaces the static bitboxbase.service configuration of Avahi mDNS daemon with a
  template that creates the static file including the version information on build.

* check if Mender is enabled before initializing /etc/mender

Because:

* the startup script fails on a non-mender-image if it tries to
  initialize the mender configuration

This commit:

* adds a check if the file /etc/mender/mender.conf (added to the image
  by the mender-convert process) exists and only then initializes
  Mender

* bitcoind: remove .onion seednodes

https://github.com/shiftdevices/bitbox-base-internal/issues/180

Because:

* The onion seednodes in bitcoin.conf don't really do anything as a
  current list of proven nodes is built in to bitcoind, and updated
  with every release.
* We remove these seednodes as they don't add value and could raise
  legitimate questions about connectivity handling.

This commit:

* remove 'seednode' entries in 'bitcoin.conf.template'

* mender: add update from flashdrive option

Because:

* System updates were only possible from a public GitHub release. For
  testing and possibly also offline updating it's helpful to provide
  the update image from a local usb flashdrive.

* The update image is expected as '/update.base' on the flashdrive
  and needs to be signed like an official OTA update image.

This commit:

* allows usage of 'bbb-config.sh mender-update install flashdrive'

* add comment divider "generic functions"

Because:

* to separate generic includes / functions from local functions used
  only in that specifc script

This commit:

* adds only comment separators

* remove trailing whitespace and fix EOF linebreaks

Because:

* Neat code is good code.

This commit:

* removes unnecessary trailing whitespace

* prometheus-bitcoin: fix typo for tor address

* bbb-cmd: remove 'shutdown' and 'restart' commands

https://github.com/shiftdevices/bitbox-base-internal/issues/300

Because:

* It is not possible to call the commands 'shutdown' or 'restart'
  asynchronously from the Middleware to get the return code.
* These commands are now run directly from the Middleware.

This commit:

* removes the 'shutdown' and 'restart' commands from the script

* scripts: no error output when overlayroot disabled

https://github.com/shiftdevices/bitbox-base-internal/issues/296

Because:

* The scripts need to detect if the readonly overlayroot filesystem
  is active or not.
* This can be accomplished by parsing /etc/overlayroot.local.conf
* On images built without overlayrootfs, this file is missing,
  resulting in an unnecessary error output

This commit:

* adds a check whether /etc/overlayroot.local.conf is present

+++ armbian/base/scripts/systemd-startup-checks.sh
@@ -9,6 +9,8 @@
 
 set -eu
 
+# --- generic functions --------------------------------------------------------
+
 # include function: exec_overlayroot()
 source /opt/shift/scripts/include/exec_overlayroot.sh.inc
 
@@ -28,13 +30,13 @@ if ! grep -q '/mnt/ssd ' /etc/fstab ; then
 
     else
         ## if no valid partition present, is image configured for autosetup of SSD?
-        
+
         if ! mountpoint /mnt/ssd -q && [ -f /opt/shift/config/.autosetup-ssd ]; then
             # run ssd autosetup, and disable it afterwards on success
             if /opt/shift/scripts/autosetup-ssd.sh format auto --assume-yes
             then
                 echo "INFO: autosetup-ssd.sh successfully executed"
-                /opt/shift/scripts/bbb-config.sh disable autosetup_ssd 
+                /opt/shift/scripts/bbb-config.sh disable autosetup_ssd
             else
                 echo "ERR: autosetup-ssd.sh failed"
             fi
@@ -58,7 +60,7 @@ fi
 sudo mount -a
 
 ## abort check if SSD mount is not successful
-if ! mountpoint /mnt/ssd -q; then 
+if ! mountpoint /mnt/ssd -q; then
     echo "Mounting of SSD failed"
     errorExit SSD_NOT_MOUNTED
 fi
@@ -66,7 +68,7 @@ fi
 # Folders & permissions
 # ------------------------------------------------------------------------------
 ## create missing directories & always set correct owner
-## access control lists (setfacl) are used to control permissions of newly created files 
+## access control lists (setfacl) are used to control permissions of newly created files
 chown bitcoin:system /mnt/ssd/
 
 ## bitcoind data storage
@@ -160,7 +162,7 @@ swapon /mnt/ssd/swapfile || true
 # Firmware updates
 # ------------------------------------------------------------------------------
 # initialize mender configuration
-if ! grep -q '/shift/' /etc/mender/mender.conf ; then
+if [[ -f /etc/mender/mender.conf ]] && ! grep -q '/shift/' /etc/mender/mender.conf ; then
     exec_overlayroot all-layers 'rm -f /etc/mender/mender.* /etc/mender/server.crt || true'
     exec_overlayroot all-layers 'cp -f /opt/shift/config/mender/mender.conf /etc/mender/'
 fi
